requirements.txt
asgiref==3.6.0
boto3==1.26.118
botocore==1.29.118
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
Django==4.2
django-browser-reload==1.8.0
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.1
django-resized==1.0.2
django-storages==1.13.2
django-tailwind==3.5.0
djangorestframework==3.14.0
gcloud==0.18.3
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
jmespath==1.0.1
jwcrypto==1.4.2
oauth2client==4.1.3
Pillow==9.5.0
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase4==4.6.0
python-dateutil==2.8.2
python-decouple==3.8
python-jwt==4.0.0
pytz==2023.3
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
wrapt==1.15.0
